Пытаюсь избавиться от Context в ViewModel

+++ app/src/main/java/com/example/cft/model/Bin.kt
@@ -24,8 +24,8 @@ data class BinCountry(
     val name: String? = null,
     val emoji: String? = null,
     val currency: String? = null,
-    val latitude: Int? = null,
-    val longitude: Int? = null
+    val latitude: Float? = null,
+    val longitude: Float? = null
 )
 
 @kotlinx.serialization.Serializable
